requirements.txt
# Production dependencies
python-telegram-bot==20.7
gspread==6.0.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1  
google-auth-httplib2==0.2.0
google-api-python-client==2.109.0
python-dotenv==1.0.0
pytz==2023.3
flask==3.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
